requirements.txt
google-adk
kalshi_python_sync
dotenv
google-genai